New document: Working on GitHub
This is a stub to be fleshed out very soon.

The intention is for it to contain generic rules for how we work with GitHub across all our projects.

It shouldn't contain tutorials (although it could link to them) but instead define our processes for
forking projects, approving work, etc.