Rename "solutions" to "project-structure"
"Solutions" is too general a word to really mean anything. So I've
renamed it to "project structure", which I hope conveys a bit
more meaning.

I've also moved "ports" and "the-run-script" here from
"local-development". This reduces the number of top-level categories,
which has been balooning a bit, and I hope is more useful logical
grouping: People developing on our applications locally are probably
interested in the whole structure of the project, not just the
parts that are exclusively for local development.